First try of mocking the login bean
I'm sure this can be done better. Why don't we learn this shit in class?
Why do we always have to figure out everything on our own, have it
rubber-stamped in the peer reviews because no one else has a clue either,
only to get berated weeks later since everything was wrong from the start.
Fuck this shit. If this is the future of software development, I'm going
to be a truck driver or something.