TESB416-165: Hardcode webdriver versions
By hardcoding the versions, all but the first request
will be satisfied by the local cache, so that there
is practically no chance for GitHub to ever lock us
out again.

Downside: Every few weeks the versions need to be updated
to restore compatibility with later browser versions.

In the long run, we should try to fix the library
to cache the latest versions for a day or so.