TESB416-232: Kill local test server after timeout
In contrast to the pipe-based solution, this should
work on windows, too.

My windows installation unfortunately chokes on
the cucumber tests. We'll see what jenkins says.

TODO: This timeout solution can only be considered
a temporary fix. And it already has a known bug:
When running the tests on the command line the test
suite will not complete until the api times out.